Begin moving ListVisualizer to jsx
