+++ package.json
@@ -48,6 +48,8 @@
     "flexboxgrid": "^6.3.1",
     "flexboxgrid-helpers": "^1.1.3",
     "js-slang": "^0.4.73",
+    "konva": "^7.2.5",
+    "konva-node": "^0.11.2",
     "lodash": "^4.17.20",
     "lz-string": "^1.4.4",
     "moment": "^2.29.1",
@@ -62,6 +64,7 @@
     "react-draggable": "^4.4.3",
     "react-dropzone": "^11.2.4",
     "react-hotkeys": "^2.0.0",
+    "react-konva": "^17.0.2-0",
     "react-mde": "^11.0.6",
     "react-redux": "^7.2.2",
     "react-responsive": "^8.2.0",
